// src/ts/enums/style_constant_enums.ts
enum COLOR {
  black = "#000000",
  blue = "#0089ED",
  blue1 = "#0089ED",
  white = "#FFFFFF",
  white1 = "#C1C7D0",
  green1 = "#01A885",
  neutral = "#091E42",
  neutral80 = "#091e426b",
  neutral90 = "#091E42",
  neutral100 = "#101828",
  neutral110 = "#7A869A",
  blue50 = "#f2f7fd",
}

enum FONT_SIZE {
  default = "12px",
  fs1 = "14px",
  fs2 = "16px",
  fs3 = "18px",
  fs4 = "20px",
  fs5 = "22px",
  fs6 = "24px",
  fs7 = "26px",
}

enum FONT_WEIGHT {
  normal = 400,
  medium = 500,
  semiBold = 600,
  bold = 700,
}

enum LINE_HEIGHT {
  normal = 1,
  lh1 = 1.3,
  lh2 = 1.5,
}

export { LINE_HEIGHT, COLOR, FONT_SIZE, FONT_WEIGHT };